requirements.txt
fastapi[standard]
python-dotenv
supabase
sqlalchemy
psycopg2
